Don't retry requests or show error toast for 404 (i.e. for person search)
